requirements.txt
pandas
python-dotenv
requests
beautifulsoup4
lxml
html5lib